requirements.txt
feedparser>=6.0.8
requests>=2.28.0
python-dateutil>=2.8.2
playwright>=1.42.0
playwright-stealth>=1.0.6
supabase>=2.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0.0
schedule>=1.2.0
beautifulsoup4>=4.9.3
PyMuPDF>=1.23.0
newspaper3k>=0.2.8
httpx>=0.25.0
groq>=0.5.0
pytest>=7.0.0